refactor: MultiCheckLayout과 SingleCheckLayout 하나로 병합
